+++ src/services/axios/getProposal/mockProposalBackend.tsx
@@ -1,4 +1,5 @@
 import { ProposalBackend } from '../../../utils/types/proposal';
+import { RA_TYPE_ICRS } from '@/utils/constants';
 
 export const MockProposalBackend: ProposalBackend = {
   prsl_id: 'prsl-t0001-20250613-00002',
@@ -38,12 +39,11 @@ export const MockProposalBackend: ProposalBackend = {
           ]
         },
         reference_coordinate: {
-          kind: 'equatorial',
-          ra: '00:00:00.0',
-          dec: '00:00:00.0',
-          reference_frame: 'icrs',
-          epoch: 2000,
-          unit: ['hourangle', 'deg']
+          kind: RA_TYPE_ICRS.label,
+          ra_str: '00:00:00.0',
+          dec_str: '00:00:00.0',
+          reference_frame: RA_TYPE_ICRS.label,
+          epoch: 2000
         },
         radial_velocity: {
           quantity: {
@@ -69,12 +69,11 @@ export const MockProposalBackend: ProposalBackend = {
           ]
         },
         reference_coordinate: {
-          kind: 'equatorial',
-          ra: '05:34:30.900',
-          dec: '+22:00:53.000',
-          reference_frame: 'icrs',
-          epoch: 2000,
-          unit: ['hourangle', 'deg']
+          kind: RA_TYPE_ICRS.label,
+          ra_str: '05:34:30.900',
+          dec_str: '+22:00:53.000',
+          reference_frame: RA_TYPE_ICRS.label,
+          epoch: 2000
         },
         radial_velocity: {
           quantity: {
@@ -100,12 +99,11 @@ export const MockProposalBackend: ProposalBackend = {
           ]
         },
         reference_coordinate: {
-          kind: 'equatorial',
-          ra: '21:33:27.0200',
-          dec: '-00:49:23.700',
-          reference_frame: 'icrs',
-          epoch: 2000,
-          unit: ['hourangle', 'deg']
+          kind: RA_TYPE_ICRS.label,
+          ra_str: '21:33:27.0200',
+          dec_str: '-00:49:23.700',
+          reference_frame: RA_TYPE_ICRS.label,
+          epoch: 2000
         },
         radial_velocity: {
           quantity: {
@@ -255,12 +253,11 @@ export const MockProposalBackendZoom: ProposalBackend = {
           ]
         },
         reference_coordinate: {
-          kind: 'equatorial',
-          ra: '21:33:27.0200',
-          dec: '-00:49:23.700',
-          reference_frame: 'icrs',
-          epoch: 2000,
-          unit: ['hourangle', 'deg']
+          kind: RA_TYPE_ICRS.label,
+          ra_str: '21:33:27.0200',
+          dec_str: '-00:49:23.700',
+          reference_frame: RA_TYPE_ICRS.label,
+          epoch: 2000
         },
         radial_velocity: {
           quantity: {