+++ CHANGELOG.md
@@ -8,6 +8,8 @@ UNRELEASED
 ------------
 * Updated layout of tables in Overview Page
 
+* Changed target types and mappings to use ADR-63 format
+
 3.0.0
 
 -------